requirements.txt
torch
torchvision
pandas
scikit-image
tqdm
Pillow
python-math